Tornou a classe pessoa e funcionario abstratas.

+++ fundamentos_oo/src/main/kotlin/one/digitalinnovation/digionebank/Funcionario.kt
@@ -1,9 +1,11 @@
 package one.digitalinnovation.digionebank
 
+import java.math.BigDecimal
+
 abstract class Funcionario(
     nome: String,
     cpf: String,
-    val salario: Double,
+    val salario: BigDecimal,
 ) : Pessoa(nome, cpf) {
     protected abstract fun calculoAuxilio(): Double
